Move view functions into own file
